pyproject.toml
[project]
name = "etl-mi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=1.6.17",
    "kagglehub>=0.3.4",
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "pyspark>=3.5.3",
]
